backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Face Recognition
face-recognition==1.3.0
dlib==19.24.2
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Vector Database
qdrant-client==1.7.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2
